Add std::thread::available_concurrency
